Use vite for js bundling
