+++ website/js/main.js
@@ -1,0 +1,2 @@
+import './search.js'
+import './toggle-color-scheme.js'
