+++ website/js/toggle-color-scheme.js
@@ -1,0 +1,2 @@
+document.getElementById("light-switch-navbar").onclick = window.toggleColorScheme;
+document.getElementById("light-switch-mobile-nav").onclick = window.toggleColorScheme;
