+++ website/justfile
@@ -1,17 +1,20 @@
-build: docs playground-release
+build $BUILD_MODE="production": npm docs playground-release
+  npx vite build
   zola build
 
-build-debug: docs playground
+build-debug: npm docs playground
+  npx vite build
   zola build
 
 serve *args: docs playground
-  zola serve --open --fast {{args}}
+  zola serve --open --fast {{args}} & npx vite serve & wait
 
-serve-release: docs playground-release
-  zola serve --open --fast
+serve-release $BUILD_MODE="production": docs playground-release
+  zola serve --open --fast & npx vite serve & wait
 
 clean:
   cargo clean
+  rm -rf node_modules
   rm -rf public
 
 deploy:
@@ -34,6 +37,9 @@ docs: search-index
 
 docs-watch:
   cargo watch -w ../modules/koto -w xtask -x "xtask docs && xtask search-index"
+
+npm:
+  npm install
 
 playground:
   cd playground && trunk build
@@ -71,6 +77,9 @@ version-snapshot version: docs playground-release
   git add static/play-{{version}}
   git add templates/data.toml
   git commit -m "Snapshot of version {{version}}"
+
+vite *args:
+  npx vite build {{args}}
 
 watch command:
   cargo watch -s "just {{command}}"
