+++ website/static/.gitignore
@@ -1,2 +1,3 @@
+bundle
 play
 search-index-next.json
