+++ website/templates/base.html
@@ -1,4 +1,10 @@
 {% import "components.html" as components %}
+
+{% set build_mode = get_env(name="BUILD_MODE", default="development") %}
+{% if build_mode == "production" %}
+  {% set manifest = load_data(path="/bundle/.vite/manifest.json") %}
+{% endif %}
+
 {% set data = load_data(path="templates/data.toml") %}
 
 <!DOCTYPE html>
@@ -18,12 +24,15 @@
       <link rel="alternate" type="application/atom+xml" title="Feed" href="{{ get_url(path="atom.xml", trailing_slash=false) }}">
     {% endblock %}
 
-    <script src="https://cdn.jsdelivr.net/npm/fuse.js@7.1.0/dist/fuse.min.js"></script>
     <script src="/uikit.min.js"></script>
     <script src="/uikit-icons.min.js"></script>
 
-    <script src="/color-scheme.js"></script>
-    <script defer src="/search.js"></script>
+    {% if build_mode == "production" %}
+      <script src="/bundle/{{ manifest["js/color-scheme.js"]["file"] }}"></script>
+    {% else %}
+      <script type="module" src="http://localhost:5173/@vite/client"></script>
+      <script src="http://localhost:5173/js/color-scheme.js"></script>
+    {% endif %}
 
     {% block head %} {% endblock head %}
   </head>
@@ -37,5 +46,10 @@
     {% block mobile_nav %} {% endblock %}
     {% include "mobile-nav-end.html" %}
 
+    {% if build_mode == "production" %}
+      <script type="module" src="/bundle/{{ manifest["js/main.js"]["file"] }}"></script>
+    {% else %}
+      <script type="module" src="http://localhost:5173/js/main.js"></script>
+    {% endif %}
   </body>
 </html>
