+++ website/templates/header.html
@@ -49,8 +49,8 @@
         </li>
 
         <li>
-          <a class="uk-navbar-item light-switch" onclick="toggleColorScheme();">
-            <img class="light-switch-icon" id="light-switch-icon-navbar" />
+          <a class="uk-navbar-item light-switch" id="light-switch-navbar">
+            <img class="light-switch-icon" id="light-switch-navbar-icon" />
           </a>
         </li>
