+++ website/templates/mobile-nav-start.html
@@ -9,8 +9,8 @@
       <a class="uk-icon-link" uk-icon="discord" href="https://discord.gg/JeV8RuK4CT"></a>
       <a class="uk-icon-link" uk-icon="github" href="https://github.com/koto-lang/koto"></a>
 
-      <a class="uk-icon-link light-switch" uk-icon="" onclick="toggleColorScheme();">
-        <img class="light-switch-icon" id="light-switch-icon-mobile-nav"/>
+      <a class="uk-icon-link light-switch" uk-icon="" id="light-switch-mobile-nav">
+        <img class="light-switch-icon" id="light-switch-mobile-nav-icon"/>
       </a>
     </div>
